requirements-dev.txt
-r requirements.txt
pytest-homeassistant-custom-component==0.13.153
pytest-asyncio
ruff
